test: include test for accrual arrears
