+++ hrms/payroll/doctype/salary_slip/test_salary_slip.py
@@ -1821,77 +1821,6 @@ class TestSalarySlip(IntegrationTestCase):
 
 		self.assertEqual(salary_slip.total_income_tax, total_income_tax)
 
-	@change_settings(
-		"Payroll Settings",
-		{"payroll_based_on": "Leave"},
-	)
-	def test_payroll_correction(self):
-		from hrms.payroll.doctype.salary_structure.test_salary_structure import make_salary_structure
-
-		emp = make_employee(
-			"test_payroll_correction@salary.com",
-			company="_Test Company",
-			date_of_joining="2021-01-01",
-		)
-		payroll_period = frappe.get_last_doc("Payroll Period", filters={"company": "_Test Company"})
-		salary_structure_doc = make_salary_structure(
-			"Test Payroll Correction",
-			"Monthly",
-			company="_Test Company",
-			employee=emp,
-			payroll_period=payroll_period,
-			test_arrear=True,
-			base=65000,
-		)
-
-		leave_application = frappe.get_doc(
-			dict(
-				doctype="Leave Application",
-				employee=emp,
-				leave_type="Leave Without Pay",
-				from_date=payroll_period.start_date,
-				to_date=payroll_period.start_date,
-				company="_Test Company",
-				status="Approved",
-				leave_approver="test@example.com",
-			)
-		).insert()
-		leave_application.submit()
-
-		salary_slip = make_salary_slip(
-			salary_structure_doc.name, employee=emp, posting_date=payroll_period.start_date
-		)
-		salary_slip.save()
-		salary_slip.submit()
-
-		arrear_doc = frappe.get_doc(
-			dict(
-				doctype="Payroll Correction",
-				employee=emp,
-				payroll_period=payroll_period.name,
-				payroll_date=add_days(payroll_period.start_date, 32),  # next month
-				company="_Test Company",
-				days_to_reverse=1,
-				month_for_lwp_reversal=calendar.month_name[payroll_period.start_date.month],
-				salary_slip_reference=salary_slip.name,
-				working_days=27,
-				lwp_days=1,
-				total_lwp_applied=1,
-			)
-		).save()
-		arrear_doc.submit()
-
-		basic_salary_arrear = (65000 / 27) * 1
-		self.assertEqual(arrear_doc.get("earning_arrears")[0].amount, flt(basic_salary_arrear, 2))
-		self.assertTrue(
-			frappe.db.exists(
-				"Additional Salary",
-				{
-					"ref_docname": arrear_doc.name,
-				},
-			)
-		)
-
 
 class TestSalarySlipSafeEval(IntegrationTestCase):
 	def test_safe_eval_for_salary_slip(self):
@@ -2189,7 +2118,9 @@ def make_deduction_salary_component(setup=False, test_tax=False, company_list=No
 	return data
 
 
-def make_employee_benefit_earning_components(setup=False, test_tax=False, company_list=None):
+def make_employee_benefit_earning_components(
+	setup=False, test_tax=False, company_list=None, test_arrear=False
+):
 	if setup:
 		data = [
 			{
@@ -2219,6 +2150,14 @@ def make_employee_benefit_earning_components(setup=False, test_tax=False, compan
 				"accrual_component": 1,
 			},
 		]
+
+		if test_arrear:
+			data[1].update(
+				{
+					"arrear_component": 1,
+					"depends_on_payment_days": 1,
+				}
+			)
 
 		make_salary_component(data, test_tax, company_list)
